alcove_env: call per file
